05_Authentication && 1:N model relation

+++ posts/views.py
@@ -1,4 +1,5 @@
 from django.shortcuts import render, redirect, get_list_or_404, get_object_or_404
+from django.contrib.auth.decorators import login_required
 from .form import PostForm, ImageForm
 from .models import Post, Image
 
@@ -9,12 +10,15 @@ def list(request):
         'posts': posts,
     }
     return render(request, 'posts/list.html', context)
-    
+
+@login_required
 def create(request):
     if request.method == 'POST':
         post_form = PostForm(request.POST)
         if post_form.is_valid():
-            post = post_form.save()
+            post = post_form.save(commit=False)
+            post.user = request.user
+            post.save()
             for image in request.FILES.getlist('file'):
                 request.FILES['file'] = image
                 image_form = ImageForm(files=request.FILES)
@@ -32,26 +36,31 @@ def create(request):
     }
     return render(request, 'posts/form.html', context)
 
+@login_required
 def update(request, post_pk):
     post = get_object_or_404(Post, pk=post_pk)
-    if request.method == 'POST':
-        post_form = PostForm(request.POST, instance=post)
-        if post_form.is_valid():
-            post_form.save()
-            return redirect('posts:list')
+    if post.user == request.user:
+        if request.method == 'POST':
+            post_form = PostForm(request.POST, instance=post)
+            if post_form.is_valid():
+                post_form.save()
+                return redirect('posts:list')
+        else:
+            post_form = PostForm(instance=post)
     else:
-        post_form = PostForm(instance=post)
+        return redirect('posts:list')
     
     context = {
         'post_form': post_form,
         'post': post,
     }
     return render(request, 'posts/form.html', context)
-    
+
+@login_required
 def delete(request, post_pk):
     post = get_object_or_404(Post, pk=post_pk)
-    if request.method == "POST":
-        post.delete()
-        return redirect('posts:list')
-    else:
-        return redirect('posts:list')
+    if post.user == request.user:
+        if request.method == "POST":
+            post.delete()
+            return redirect('posts:list')
+    return redirect('posts:list')
